Add files via upload
These files contain the code for the game of life using Conways rules. The Methods file contains methods that are helpful to the main and unrelated to the game of life. The main program allows to test the game. use ./life to figure out the usage. Pausing has been implemented but is not triggered by anything in the program.